AI-Powered-Cloud-Monitoring-and-Auto-Healing-System-main/requirements.txt
psutil
pyyaml
pandas
scikit-learn
joblib
proxmoxer
requests
